Moved test classes to proper packages
